add validation + test

+++ post-writer/src/main/java/pw/feed/postwriter/model/post/PostRepository.java
@@ -7,12 +7,12 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Set;
 
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
     @Query("""
-            SELECT p 
-            FROM Post p 
-            WHERE p.user.id IN :userIds 
-            OR p.group.id IN :groupIds 
+            SELECT p
+            FROM Post p
+            WHERE p.user.id IN :userIds
+            OR p.group.id IN :groupIds
             ORDER BY p.createdAt DESC""")
     Set<Post> findPostsByFollowedUsersOrGroups(@Param("userIds") List<Integer> userIds, @Param("groupIds") List<Integer> groupIds);
 }
